+++ src/randomizer.rs
@@ -1,5 +1,7 @@
 #![allow(dead_code)]
 
+use crate::character::StatusEffect;
+use crate::game::battle::AttackType;
 use crate::location;
 use rand::Rng;
 use std::cmp::max;
@@ -19,13 +21,18 @@ pub trait Randomizer {
 
     fn damage(&self, value: i32) -> i32;
 
-    fn is_critical(&self) -> bool;
-
-    fn is_miss(&self, attacker_speed: i32, receiver_speed: i32) -> bool;
+    fn attack_type(
+        &self,
+        produced_status: Option<(StatusEffect, u32)>,
+        attacker_speed: i32,
+        receiver_speed: i32,
+    ) -> AttackType;
 
     fn gold_gained(&self, base: i32) -> i32;
 
     fn stat_increase(&self, increase: i32) -> i32;
+
+    fn range(&self, max: i32) -> i32;
 }
 
 #[cfg(not(test))]
@@ -82,19 +89,23 @@ impl Randomizer for DefaultRandomizer {
         max(1, rng.gen_range(min_val..=max_val))
     }
 
-    fn is_critical(&self) -> bool {
-        let mut rng = rand::thread_rng();
-        rng.gen_ratio(1, 20)
-    }
-
-    fn is_miss(&self, attacker_speed: i32, receiver_speed: i32) -> bool {
-        if receiver_speed > attacker_speed {
-            let ratio = receiver_speed / attacker_speed;
-            let ratio = max(1, 5 - ratio) as u32;
-            let mut rng = rand::thread_rng();
-            return rng.gen_ratio(1, ratio);
+    fn attack_type(
+        &self,
+        produced_status: Option<(StatusEffect, u32)>,
+        attacker_speed: i32,
+        receiver_speed: i32,
+    ) -> AttackType {
+        // TODO this could maybe flattened to a single random call
+        // instead of trial and error
+        if is_miss(attacker_speed, receiver_speed) {
+            AttackType::Miss
+        } else if is_critical() {
+            AttackType::Critical
+        } else if let Some((status, ratio)) = produced_status {
+            status_attack(status, ratio)
+        } else {
+            AttackType::Regular
         }
-        false
     }
 
     fn gold_gained(&self, base: i32) -> i32 {
@@ -110,6 +121,35 @@ impl Randomizer for DefaultRandomizer {
 
         let mut rng = rand::thread_rng();
         rng.gen_range(min_value..=max_value)
+    }
+
+    fn range(&self, max: i32) -> i32 {
+        let mut rng = rand::thread_rng();
+        rng.gen_range(0..max)
+    }
+}
+
+fn is_critical() -> bool {
+    let mut rng = rand::thread_rng();
+    rng.gen_ratio(1, 20)
+}
+
+fn is_miss(attacker_speed: i32, receiver_speed: i32) -> bool {
+    if receiver_speed > attacker_speed {
+        let ratio = receiver_speed / attacker_speed;
+        let ratio = max(1, 5 - ratio) as u32;
+        let mut rng = rand::thread_rng();
+        return rng.gen_ratio(1, ratio);
+    }
+    false
+}
+
+fn status_attack(status: StatusEffect, ratio: u32) -> AttackType {
+    let mut rng = rand::thread_rng();
+    if rng.gen_ratio(1, ratio) {
+        AttackType::Effect(status)
+    } else {
+        AttackType::Regular
     }
 }
 
@@ -138,12 +178,13 @@ impl Randomizer for TestRandomizer {
         value
     }
 
-    fn is_critical(&self) -> bool {
-        false
-    }
-
-    fn is_miss(&self, _attacker_speed: i32, _receiver_speed: i32) -> bool {
-        false
+    fn attack_type(
+        &self,
+        _produced_status: Option<(StatusEffect, u32)>,
+        _attacker_speed: i32,
+        _receiver_speed: i32,
+    ) -> AttackType {
+        AttackType::Regular
     }
 
     fn gold_gained(&self, base: i32) -> i32 {
@@ -152,6 +193,10 @@ impl Randomizer for TestRandomizer {
 
     fn stat_increase(&self, increase: i32) -> i32 {
         increase
+    }
+
+    fn range(&self, max: i32) -> i32 {
+        max
     }
 }